Renamed everything from myads to vault
