+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-addopts = --cov=myads_service --cov-report=term-missing
-testpaths = myads_service/tests
+addopts = --cov=vault_service --cov-report=term-missing
+testpaths = vault_service/tests
