+++ vault_service/tests/test_bumblebee.py
@@ -9,8 +9,8 @@ project_home = os.path.abspath(os.path.join(os.path.dirname(__file__), '../../')
 if project_home not in sys.path:
     sys.path.insert(0, project_home)
 
-from myads_service import app
-from myads_service.models import Query, Institute, Library, Base
+from vault_service import app
+from vault_service.models import Query, Institute, Library, Base
 
 class TestSite(TestCase):
     '''Tests that each route is an http response'''
@@ -23,7 +23,7 @@ class TestSite(TestCase):
                'TESTING': True,
                'PROPAGATE_EXCEPTIONS': True,
                'TRAP_BAD_REQUEST_ERRORS': True,
-               'MYADS_BUMBLEBEE_OPTIONS': {'foo': 'bar'}
+               'VAULT_BUMBLEBEE_OPTIONS': {'foo': 'bar'}
             })
         a.db.create_all()
         return a
@@ -57,7 +57,7 @@ class TestOpenURL(TestCase):
                'TESTING': True,
                'PROPAGATE_EXCEPTIONS': True,
                'TRAP_BAD_REQUEST_ERRORS': True,
-               'MYADS_BUMBLEBEE_OPTIONS': {'foo': 'bar'}
+               'VAULT_BUMBLEBEE_OPTIONS': {'foo': 'bar'}
             })
         Base.query = a.db.session.query_property()
         Base.metadata.create_all(bind=a.db.engine)
