+++ vault_service/views/query_as_monument.py
